fix(arch): relocate banner/progression/scheduling types into domain to satisfy layer rules
