+++ src/main/kotlin/net/lumalyte/lg/application/services/PenaltyService.kt
@@ -1,6 +1,7 @@
 package net.lumalyte.lg.application.services
 
 import net.lumalyte.lg.application.persistence.PenaltyRepository
+import net.lumalyte.lg.domain.values.ExperienceSource
 import net.lumalyte.lg.config.StrikesConfig
 import net.lumalyte.lg.domain.entities.Guild
 import net.lumalyte.lg.domain.entities.GuildPenalty
