+++ src/main/kotlin/net/lumalyte/lg/infrastructure/services/scheduling/TaskBukkit.kt
@@ -1,6 +1,6 @@
 package net.lumalyte.lg.infrastructure.services.scheduling
 
-import net.lumalyte.lg.application.services.scheduling.Task
+import net.lumalyte.lg.domain.services.Task
 import org.bukkit.scheduler.BukkitRunnable
 
 class TaskBukkit(private val runnable: BukkitRunnable) : Task {
